requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
google-generativeai>=0.5.0
sentence-transformers>=2.2.2
langchain>=0.2.0
langchain-text-splitters>=0.0.2
langchain-core>=0.2.0
langchain-community>=0.1.0
langchain-openai>=0.1.0
faiss-cpu>=1.13.0
openai>=1.3.0
